Intial FB update to new API
